requirements.txt
python-dotenv
requests
pydantic
chromadb
fastapi
uvicorn
openai
